Development3.0/TeevraServer/extensions/generic/connectors/filesplitter/src/main/java/com/headstrong/fusion/connector/filesplitter/FileSplitterComponent.java
/*
 * The information in this document is subject to change without notice and 
 * does not represent a commitment by Headstrong Corporation. The software 
 * and/or databases described in this document are furnished under a license 
 * agreement and may be used or copied only in accordance with the terms of 
 * the agreement. 
 * 
 * Copyright © 2008 Headstrong Corporation
 * All rights reserved.
 * 
 * $Id: FileSplitterComponent.java
 * $Revision: 
 * $Author: ViswanathP
 * $DateTime: Mar 2, 2010 
 */

package com.headstrong.fusion.connector.filesplitter;

import com.headstrong.fusion.commons.BindingConfig;
import com.headstrong.fusion.commons.MessageExchange;
import com.headstrong.fusion.commons.component.FusionComponent;
import com.headstrong.fusion.commons.component.FusionEndpoint;

/**
 * File SPlitter Component
 */
public class FileSplitterComponent implements FusionComponent {

	/**
	 * Default Constructor
	 */
	public FileSplitterComponent() {
	}

	/*
	 * (non-Javadoc)
	 * 
	 * @see com.headstrong.fusion.commons.component.FusionComponent#createEndpoint(com.headstrong.fusion.commons.BindingConfig)
	 */
	@Override
	public FusionEndpoint<MessageExchange> createEndpoint(
			BindingConfig bindingConfig) throws Exception {
		FileSplitterEndpoint fileSplitterEndpoint = new FileSplitterEndpoint(
				bindingConfig.getProperties());
		return fileSplitterEndpoint;
	}
}
